Livraison Lot1b

+++ src/UcaBundle/Controller/UcaGest/Activite/FormatActiviteController.php
@@ -25,7 +25,7 @@ class FormatActiviteController extends Controller
 {
 
     /**
-     * @Route("/Voir/{id}", name="FormatActiviteVoir")
+     * @Route("/Voir/{id}", name="UcaGest_FormatActiviteVoir")
      * @Method("GET")
      * @Isgranted("ROLE_GESTION_FORMAT_ACTIVITE_LECTURE")
      */
@@ -46,9 +46,34 @@ class FormatActiviteController extends Controller
         return $this->render('@Uca/UcaGest/Activite/FormatActivite/Voir.html.twig', $twigConfig);
     }
 
+    public function redirectAfterCommit($form, $formatActivite)
+    {
+        if ($form->get('previsualiser')->isClicked()) {
+            return $this->redirectToRoute('UcaWeb_FormatActiviteDetail', [
+                'idCa' => $formatActivite->getActivite()->getClasseActivite()->getId(),
+                'idA' => $formatActivite->getActivite()->getId(),
+                'id' => $formatActivite->getId(),
+                'previsualisation' => 'on',
+                'urlRetourPrevisualisation' => $this->generateUrl('UcaGest_FormatActiviteModifier', [
+                    'idActivite' => $formatActivite->getActivite()->getId(),
+                    'id' => $formatActivite->getId(),
+                    'previsualisation' => 'off'
+                ])
+            ]);
+        } else {
+            return $this->redirectToRoute(
+                'UcaGest_FormatActiviteVoir',
+                [
+                    'idActivite' => $formatActivite->getActivite()->getId(),
+                    'id' => $formatActivite->getId()
+                ]
+            );
+        }
+    }
+
     /**
      * @Security("is_granted('ROLE_GESTION_FORMAT_ACTIVITE_ECRITURE') and is_granted('ROLE_GESTION_ACTIVITE_ECRITURE')")
-     * @Route("/Ajouter", name="FormatActiviteAjouter")
+     * @Route("/Ajouter", name="UcaGest_FormatActiviteAjouter")
      */
     public function ajouterAction(Request $request, $idActivite)
     {
@@ -60,13 +85,18 @@ class FormatActiviteController extends Controller
             $item = new $className();
             $item->setEstEncadre(false);
             $item->setEstPayant(false);
-            $item->setActivite($em->getReference(Activite::class, $idActivite));
-            $item->setDateDebutEffective(new \Datetime('now'));
-            $item->setDateFinEffective(new \Datetime('now'));
-            $item->setDateDebutInscription(new \Datetime('now'));
-            $item->setDateFinInscription(new \Datetime('now'));
-            // DTI - 0001836 Ajout de l'image de l'activité par défaut
-            $item->setImage($this->getDoctrine()->getRepository(Activite::class)->findOneById($idActivite)->getImage());
+            $today = new \Datetime();
+            $today->setTime(0, 0);
+            $item->setDateDebutPublication($today);
+            $item->setDateFinPublication($today);
+            $item->setDateDebutEffective($today);
+            $item->setDateFinEffective($today);
+            $item->setDateDebutInscription($today);
+            $item->setDateFinInscription($today);
+            $activite = $em->getRepository(Activite::class)->find($idActivite);
+            $item->setActivite($activite);
+            $item->setImage($activite->getImage());
+            $item->setLibelle($activite->getLibelle());
             if (!$request->isMethod('POST')) {
                 $profils = $this->getDoctrine()->getRepository(ProfilUtilisateur::class)->findAll();
                 foreach ($profils as $profil) {
@@ -82,12 +112,12 @@ class FormatActiviteController extends Controller
             throw new \Exception("Format <$format> d'activité non valide");
         }
         $form = $this->get('form.factory')->create($typeClassName, $item);
-
         if ($request->isMethod('POST') && $form->handleRequest($request)->isValid()) {
+            $item->verifieCoherenceDonnees();
             $em->persist($item);
             $em->flush();
             $this->get('uca.flashbag')->addActionFlashBag($item, 'Ajouter');
-            return $this->redirectToRoute('FormatActiviteVoir', array('idActivite' => $idActivite, 'id' => $item->getId()));
+            return $this->redirectAfterCommit($form, $item);
         }
         $twigConfig['FormatClassName'] = $className;
         $twigConfig['item'] = $item;
@@ -96,7 +126,7 @@ class FormatActiviteController extends Controller
     }
 
     /** 
-     * @Route("/Modifier/{id}", name="FormatActiviteModifier")
+     * @Route("/Modifier/{id}", name="UcaGest_FormatActiviteModifier")
      * @Method({"GET""POST"})
      * @Isgranted("ROLE_GESTION_FORMAT_ACTIVITE_ECRITURE")
      */
@@ -111,12 +141,10 @@ class FormatActiviteController extends Controller
 
         $form = $this->get('form.factory')->create($typeClassName, $item); //, ['sub_class' => get_class($item)]);
         if ($request->isMethod('POST') && $form->handleRequest($request)->isValid()) {
+            $item->verifieCoherenceDonnees();
             $em->flush();
             $this->get('uca.flashbag')->addActionFlashBag($item, 'Modifier');
-            return $this->redirectToRoute('FormatActiviteVoir', [
-                'idActivite' => $idActivite,
-                'id' => $item->getId()
-            ]);
+            return $this->redirectAfterCommit($form, $item);
         }
         $twigConfig['FormatClassName'] = $className;
         $twigConfig['item'] = $item;
@@ -126,18 +154,22 @@ class FormatActiviteController extends Controller
 
     /**
      * @Security("is_granted('ROLE_GESTION_FORMAT_ACTIVITE_ECRITURE') and is_granted('ROLE_GESTION_ACTIVITE_ECRITURE')")
-     * @Route("/Supprimer/{id}", name="FormatActiviteSupprimer")
+     * @Route("/Supprimer/{id}", name="UcaGest_FormatActiviteSupprimer")
      */
     public function supprimerAction(Request $request, $idActivite, FormatActivite $formatActivite)
     {
         $em = $this->getDoctrine()->getManager();
         if (!$formatActivite->getInscriptions()->isEmpty()) {
             $this->get('uca.flashbag')->addActionErrorFlashBag($formatActivite, 'Supprimer');
-            return $this->redirectToRoute("ActiviteVoir", array("id" => $idActivite));
+            return $this->redirectToRoute("UcaGest_ActiviteVoir", array("id" => $idActivite));
+        }
+        if ($formatActivite instanceof FormatAvecCreneau && !$formatActivite->getCreneaux()->isEmpty()) {
+            $this->get('uca.flashbag')->addActionErrorFlashBag($formatActivite, 'Supprimer');
+            return $this->redirectToRoute("UcaGest_ActiviteVoir", array("id" => $idActivite));
         }
         $em->remove($formatActivite);
         $em->flush();
         $this->get('uca.flashbag')->addActionFlashBag($formatActivite, 'Supprimer');
-        return $this->redirectToRoute('ActiviteVoir', array('id' => $idActivite));
+        return $this->redirectToRoute('UcaGest_ActiviteVoir', array('id' => $idActivite));
     }
 }
